requirements.txt
aniso8601==9.0.1
blinker==1.7.0
click==8.1.7
coverage==7.4.0
dnspython==2.0.0
exceptiongroup==1.2.0
Flask==3.0.0
Flask-RESTful==0.3.10
importlib-metadata==7.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
mongotransactions==1.2.5
packaging==23.2
pluggy==1.3.0
pymongo==3.11.4
pytest==7.4.4
pytest-cov==4.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
tomli==2.0.1
Werkzeug==3.0.1
zipp==3.17.0
